+++ data-raw/vignette_mctq.R
@@ -1,8 +1,6 @@
 # Source the file before running the function
 # Don't forget to uncomment the `library` functions below
 
-library(magrittr)
-
 # library(checkmate)
 # library(dplyr)
 # library(mctq)